requirements.txt
absl-py==1.0.0
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
atomicwrites==1.4.1
attrs==22.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.11.1
boto3==1.21.7
botocore==1.24.7
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
easydict==1.9
entrypoints @ file:///C:/ci/entrypoints_1649926621247/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.79.0
filelock==3.8.0
flatbuffers==2.0
fonttools==4.29.1
gast==0.4.0
gdown==4.5.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
gTTS==2.2.4
gunicorn==20.1.0
h11==0.13.0
h5py==3.6.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
ipykernel @ file:///C:/ci/ipykernel_1647000978151/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1657652657542/work
jedi @ file:///C:/ci/jedi_1644315425835/work
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jupyter-client @ file:///C:/ci/jupyter_client_1650534470695/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656285178/work
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929372/work
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect==4.8.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pluggy==1.0.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyMySQL==1.0.2
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PySocks==1.7.1
pytest==6.2.5
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2022.2.1
pywin32==302
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657616005830/work
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.1.2
scipy==1.8.0
seaborn==0.11.2
simplegeneric==0.8.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.31
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
toml==0.10.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.18.2
watchfiles==0.16.1
watchgod==0.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==10.2
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.13.3
zipp==3.7.0
